refactor(menu): 패키지 구조 변경 및 이름변경
