+++ src/main/java/kitchenpos/menus/tobe/application/TobeMenuService.java
@@ -10,9 +10,9 @@ import kitchenpos.menus.tobe.domain.menu.TobeMenu;
 import kitchenpos.menus.tobe.domain.menu.TobeMenuRepository;
 import kitchenpos.menus.tobe.domain.menugroup.TobeMenuGroup;
 import kitchenpos.menus.tobe.domain.menugroup.TobeMenuGroupRepository;
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProductQuantity;
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProducts;
+import kitchenpos.menus.tobe.domain.menu.TobeMenuProduct;
+import kitchenpos.menus.tobe.domain.menu.TobeMenuProductQuantity;
+import kitchenpos.menus.tobe.domain.menu.TobeMenuProducts;
 import kitchenpos.products.tobe.domain.TobeProduct;
 import kitchenpos.products.tobe.domain.TobeProductRepository;
 import org.springframework.stereotype.Service;
