+++ src/test/java/kitchenpos/TobeFixtures.java
@@ -7,8 +7,8 @@ import kitchenpos.menus.tobe.domain.menu.MenuPrice;
 import kitchenpos.menus.tobe.domain.menu.TobeMenu;
 import kitchenpos.menus.tobe.domain.menugroup.TobeMenuGroup;
 import kitchenpos.menus.tobe.domain.menugroup.TobeMenuGroupName;
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProductQuantity;
-import kitchenpos.menus.tobe.domain.menuproduct.TobeMenuProduct;
+import kitchenpos.menus.tobe.domain.menu.TobeMenuProductQuantity;
+import kitchenpos.menus.tobe.domain.menu.TobeMenuProduct;
 import kitchenpos.products.application.FakePurgomalumChecker;
 import kitchenpos.products.tobe.domain.ProductName;
 import kitchenpos.products.tobe.domain.ProductPrice;
